deduplication logic
added a step to the pipeline logic which uses an embeddings model to deduplicate comments *before* classification + plotting. thus, as we add steps between classification (types of commenters) <> plotting, ensure the logic reflects this step and this will be a great efficiency speedup. tbh not even much logic actually changs -- note the changes in classify_makeup.py to see what changed to get those effiency improvements